cache changes

+++ exareme-common/src/main/java/madgik/exareme/common/schema/Partition.java
@@ -16,7 +16,7 @@ public class Partition implements Serializable {
     private int pNum = 0;
     private Set<String> locations = null;
     private Set<String> partitionColumns = null;
-    private int sizeInBytes = 0;
+    private long sizeInBytes = 0;
 
     public Partition(String table, int pNum) {
         this.table = table;
@@ -49,11 +49,11 @@ public class Partition implements Serializable {
         return Collections.unmodifiableList(new ArrayList<String>(partitionColumns));
     }
     
-    public void setSize(int size) {
+    public void setSize(long size) {
         sizeInBytes = size;
     }
 
-    public int getSize() {
+    public long getSize() {
         return sizeInBytes;
     }
 }
